Fix int16 overflow in the gather_qmm sorted row bound
The NAX sorted gather kernel narrowed the per-simdgroup row and column
bounds to short before taking the min:

  const short sgp_sm = align_M ? SM : min(SM, short(max(0, (M - (y_row + tm)))));

For a row count above the int16 range the inner cast wraps negative, so
the min returns a negative bound and the tile is loaded with garbage
limits. The guard only runs when align_M is false, so both a large M and
an M that is not a multiple of BM are required to reach it.

Taking the min in int and narrowing afterwards is safe because the result
is bounded by SM. This matches fp_quantized_nax.h and every other site.

In a quantized MoE the row count is tokens * experts_per_token, so this is
reachable at long context. For Qwen3-Coder-30B-A3B (8 experts per token) it
corrupts any single prefill whose length is not a multiple of 8, past about
4k tokens.

+++ mlx/backend/metal/kernels/quantized_nax.h
@@ -1529,10 +1529,10 @@ template <
   const short tm = SM * (simd_group_id / WN);
   const short tn = SN * (simd_group_id % WN);
 
-  const short sgp_sm =
-      align_M ? SM : min(SM, short(max(0, (M - (y_row + tm)))));
-  const short sgp_sn =
-      align_N ? SN : min(SN, short(max(0, (N - (y_col + tn)))));
+  // Take the min in int and narrow afterwards. Narrowing first overflows for
+  // M or N above the int16 range, which wraps the bound negative.
+  const short sgp_sm = align_M ? SM : min(int(SM), max(0, (M - (y_row + tm))));
+  const short sgp_sn = align_N ? SN : min(int(SN), max(0, (N - (y_col + tn))));
 
   const bool is_unaligned_sm = align_M ? false : (sgp_sm != SM);
   const bool is_unaligned_bn = align_N ? false : (tgp_bn != BN);
